Extracting common publish functions.
